+++ __tests__/__integration__/lib.test.js
@@ -1,4 +1,4 @@
-const Spalatum = require('../../lib/index.js');
+const spalatum = require('../../lib/index.js');
 const PrimaryFragmentException = require('../../lib/exceptions/primaryFragmentException.js');
 
 const mockPath = '../../__mocks__';
@@ -10,7 +10,7 @@ beforeEach(() => {
 describe('# Testing a request to a https fragment', () => {
   it('Calling spalatum with https fragments', async () => {
     const httpsTemplate = require(`${mockPath}/https-template.js`);
-    const renderedHttpsTemplate = await Spalatum.render(httpsTemplate);
+    const renderedHttpsTemplate = await spalatum.render(httpsTemplate);
 
     document.body.outerHTML = renderedHttpsTemplate;
     expect(document.body.outerHTML).toMatchSnapshot();
@@ -20,7 +20,7 @@ describe('# Testing a request to a https fragment', () => {
 describe('# Testing a request to a fragment that throws an error', () => {
   it('Calling spalatum with a fragment that status code returns a 404, I expect that returns fragments rendered in blank', async () => {
     const template = require(`${mockPath}/notfound-template.js`);
-    const renderedTemplate = await Spalatum.render(template);
+    const renderedTemplate = await spalatum.render(template);
 
     document.body.outerHTML = renderedTemplate;
     expect(document.body.outerHTML).toMatchSnapshot();
@@ -29,7 +29,7 @@ describe('# Testing a request to a fragment that throws an error', () => {
   it('Calling spalatum with a fragment with primary attribute that status code returns a 404, I expect that returns an error object', () => {
     const template = require(`${mockPath}/notfound-primary-template.js`);
 
-    expect(Spalatum.render(template)).rejects.toEqual(
+    expect(spalatum.render(template)).rejects.toEqual(
       new PrimaryFragmentException(
         'Spalatum can\'t render the primary fragment (https://httpbin.org/notfound/), the returned statusCode was 404.',
       ),